+++ src/pages/main/SearchPop.tsx
@@ -1,6 +1,10 @@
 import React, { useEffect, useState } from 'react'
+import axios from 'axios';
+import convert from 'xml-js';
 
 import { IoClose } from "react-icons/io5";
+// import { FaRegHospital } from "react-icons/fa";
+import { FaRegHospital } from "react-icons/fa6";
 
 const SearchPop = (props) => {
     let {institution, selected, setSearchPop} = props;
@@ -13,7 +17,27 @@ const SearchPop = (props) => {
     
     useEffect(()=>{
         console.log(value)
-    },[value])
+    },[value]);
+
+
+    const fetchData = async () => {
+      try {
+        const URL = 'http://apis.data.go.kr/openapi/service/rest/HmcSearchService/getHchkTypesHmcList';
+        const response = await axios.get(URL, {
+          params: {
+            serviceKey: process.env.REACT_APP_API_KEY,
+            numOfRows: 1,
+            pageNo: 10,
+          },
+        });
+        const searchItem = response.data.response.body.items.item;
+        console.log(searchItem);
+      } catch (err) {
+        console.log(err);
+      }
+   
+    };
+    fetchData();
 
     
   return (
@@ -29,6 +53,20 @@ const SearchPop = (props) => {
               })
             }
         </select>
+
+        <div className="hospital-wrap">
+          <div className="hospital">
+            <div className="icon">
+              {/* <FaRegHospital className='fontawesome' /> */}
+              <FaRegHospital className='fontawesome' />
+            </div>
+
+            <div className="hospital-info">
+              <p>이름</p>
+              <p>주소</p>
+            </div>            
+          </div>
+        </div>
     </div>
   )
 }
