--- src/pages/main/chungcheong_north.ts
@@ -1,5 +1,0 @@
-let chungcheongNorth: string[] = [
-    '괴산군', '단양군', '보은군', '영동군', '옥천군', '음성군', '제천시', '증평군', '진천군', '청주시' , '청주시 상당구', '청주시 서원구', '청주시 청원구', '청주시 흥덕구', '충주시'
-]
-
-export default chungcheongNorth;
