--- src/pages/main/chungcheong_south.ts
@@ -1,5 +1,0 @@
-let chungcheongSouth: string[] = [
-    '계룡시', '공주시', '금산군', '논산시', '당진시', '동남구', '보령시', '부여군', '서북구', '서산시', '서천군', '아산시', '예산군', '천안시' , '천안시 동남구', '천안시 서북구', '청양군', '태안군', '홍성군'
-]
-
-export default chungcheongSouth;
