--- src/pages/main/daegu.ts
@@ -1,5 +1,0 @@
-let daegu:string[] = [
-    '군위군', '남구', '달서구', '달성군', '동구', '북구', '서구', '수성구', '중구'
-]
-
-export default daegu;
