--- src/pages/main/daejeon.ts
@@ -1,5 +1,0 @@
-let daejeon: string[] = [
-    '대덕구', '동구', '서구', '유성구', '중구'
-]
-
-export default daejeon;
