--- src/pages/main/gwangju.ts
@@ -1,5 +1,0 @@
-let gwangju: string[] = [
-    '광산구', '남구', '동구', '북구', '서구'
-]
-
-export default gwangju;
