--- src/pages/main/gyeongsang_north.ts
@@ -1,5 +1,0 @@
-let gyongsangNorth: string[] = [
-    '경산시', '경주시', '고령군', '구미시', '김천시', '남구', '문경시', '봉화군', '북구', '상주시', '성주군', '안동시', '영덕군', '영양군', '영주시', '영천시', '예천군', '울릉군', '울진군', '의성군', '청도군', '청송군', '칠곡군', '포항시', '포항시 남구', '포항시 북구'
-]
-
-export default gyongsangNorth;
