--- src/pages/main/gyeongsang_south.ts
@@ -1,5 +1,0 @@
-let gyongsangSouth: string[] = [
-    '거제시', '거창군', '고성군', '김해시', '남해군', '마산합포구', '마산회원구', '밀양시', '사천시', '산청군', '성산구', '양산시', '의령군', '의창구', '진주시', '진해구', '창녕군', '창원시', '창원시 마산합포구', '창원시 마산회원구', '창원시 성산구', '창원시 의창구', '창원시 진해구', '통영시', '하동군', '함안군', '함양군', '합천군'
-]
-
-export default gyongsangSouth;
