--- src/pages/main/incheon.ts
@@ -1,5 +1,0 @@
-let incheon: string[] = [
-    '강화군', '계양구', '남동구', '동구', '미추홀구', '부평구', '서구', '연수구', '옹진군', '중구'
-]
-
-export default incheon;
