--- src/pages/main/jeolla_north.ts
@@ -1,5 +1,0 @@
-let jeollaNorth: string[] = [
-    '고창군', '군산시', '김제시', '남원시', '덕진구', '무주군', '순창군', '완산구', '완주군', '익산시', '임실군', '장수군', '전주시', '전주시 덕진구', '전주시 완산구', '정읍시', '진안군'
-]
-
-export default jeollaNorth;
