+++ src/pages/main/main.scss
@@ -171,16 +171,6 @@
                         @include var.r-s;    
                     }
 
-                    // &::after{
-                    //     position: fixed;
-                    //     content: '';
-                    //     bottom: 16.5rem;
-                    //     left: 47.3%;
-                    //     width: 21%;
-                    //     height: 5%;
-                    //     background: rgba(241, 245, 249,0.9);
-                    // }
-
                     div{
                         width: 100%;
                         height: 70px;
