+++ src/pages/main/search.scss
@@ -93,5 +93,28 @@
             right: 30px;
             cursor: pointer;
         }
+
+        .hospital-wrap{
+            width: 100%;
+            height: calc(100% - 40px);
+            padding-top: 30px;
+
+            .hospital{
+                width: 100%;
+                height: 89px;
+                @include var.r-m;
+                background: var.$bgColor;
+                @include var.center;
+
+                .icon{
+                    background: var.$blue;
+                }
+
+                .hospital-info{
+                    width: 60%;
+                    padding-left: 20px;
+                }
+            }
+        }
     }
 }
