--- src/pages/main/sejong.ts
@@ -1,5 +1,0 @@
-let sejong:string[] = [
-    '고운동', '금남면', '다정동', '대평동', '도담동', '반곡동', '보람동', '부강면', '새롬동', '소담동', '소정면', '아름동', '연기면', '연동면', '연서면', '장군면', '전동면', '전의면', '조치원읍', '종촌동', '한솔동', '해밀동'
-]
-
-export default sejong;
