--- src/pages/main/ulsan.ts
@@ -1,5 +1,0 @@
-let ulsan:string[] = [
-    '남구', '동구', '북구', '울주군', '중구'
-]
-
-export default ulsan;
